rename : 서비스 파일 위치 이동
