+++ match-service/src/main/java/com/ohdelivery/service/match/matching/application/service/MatchingServiceImpl.java
@@ -1,7 +1,8 @@
-package com.ohdelivery.service.match.matching.application;
+package com.ohdelivery.service.match.matching.application.service;
 
 import com.ohdelivery.common.feign.GetDeliveryResponse;
 import com.ohdelivery.service.match.common.DeliveryClientService;
+import com.ohdelivery.service.match.matching.application.MatchingEventPublisher;
 import com.ohdelivery.service.match.matching.application.dto.request.AssignRiderRequest;
 import com.ohdelivery.service.match.matching.application.dto.request.CreateMatchingRequest;
 import com.ohdelivery.service.match.matching.application.dto.response.GetMatchingResponse;
