+++ match-service/src/main/java/com/ohdelivery/service/match/matching/infrastructure/messaging/KafkaDeliveryEventConsumer.java
@@ -4,8 +4,8 @@ import static com.ohdelivery.common.kafka.Topic.CREATE_DELIVERY;
 import static com.ohdelivery.service.match.rider.infrastructure.messaging.RiderKafkaConfig.MATCH_GROUP_ID;
 
 import com.ohdelivery.common.kafka.dto.CreateDeliveryEvent;
-import com.ohdelivery.service.match.matching.application.MatchingService;
 import com.ohdelivery.service.match.matching.application.dto.request.CreateMatchingRequest;
+import com.ohdelivery.service.match.matching.application.service.MatchingService;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.kafka.annotation.KafkaListener;
